+++ library/src/main/java/net/vrallev/android/task/TaskNoCallback.java
@@ -1,0 +1,15 @@
+package net.vrallev.android.task;
+
+/**
+ * @author rwondratschek
+ */
+public abstract class TaskNoCallback extends Task<Void> {
+
+    @Override
+    protected final Void execute() {
+        executeTask();
+        return null;
+    }
+
+    protected abstract void executeTask();
+}